switchwithexpression.go
Go language Switch case statement is used to evaluate an expression and run a case that matches with  this value.

There are two variations in the  syntax of Switch statement.

Switch expression {
case value1:
       statement(s)
case value2:
       statement(s)
default:
       statement(s)
}

Where switch , case and default are the keywords. expression should evaluate to a value of type as that of the case values.

there could be multible case blocks.

the defualt block is optional.

if no case value matches the expression's value, the default block is executes.

if the following example program , we have taken a simple expression of a variable today.

based on its value we are printing what day of the week it is using switch statement.