assignment.go
Assignment operators are used to assigning a value to a variable.

The left side operand of the assignment operator is a variable and right side operand of the assignment operator is a value.

The value on the right  side must be of the same data-type of the variable on the left side otherwise on the left side otherwise the compiler will raise an error.

Different type of assignment operators are:

"=" (Simple Assignment): as simplest assignment operator name suggests , this is used to assign the value on the right to the variable in the left.

"+="(Add Assignment): This is a combination of "+" and "=" operators .This operator first adds the current value of the variable on left to the value on the right and then assigns the result to the variable on the left.

"-="(Subtract Assignment): This operator is a combination of '-' and '=' operators. This operator first adds the current value of the variable on left to the value on the right and then assigns the result to the variable on the left.

"-="(Subtract Assignment): This operator is a combination of '-' and '=' operators. This operator first subtracts the current value of the variable on left to the value on the right and then assigns the result to the variable on the left.

"*="(Multiply Assignment): This operator is a combination of '*' and '=' operators. This operator first Multiplies the current value of the variable on left to the value on the right and then assigns the result to the variable on the left.

"/=" (Division Assignment): This operator is a combination of '/' and '=' operators. This operator first divides the current value of the variable on left to the value on the right and then assigns the result to the variable on the left.

"%=" (Modulus Assignment): This operator is a combination of '%' and '=' operators. This operator first Modulo the current value of the variable on left to the value on the right and then assigns the result to the variable on the left.