lastindex.go
Go - Index of Last Occurrence of Substring in String:

To get the index of last occurrence of a substring in a string in Golang programming.

call LastIndex function of strings package, and pass the string and substring as an arguments to this function.

The syntax to get theh index of the  lastoccurrence of substring substr in a string str using LastIndex function is as follows:

strings.LastIndex(str, substr)

LastIndex function returns am integer that represents the position of last occurrence substr in str. 
we will find last index of the following "Hello" from the substring.
and find the index of last occurrence of substring.