typeassertions3.go
Here ,we can see that we get both the data and boolean value. we can use this to avoid panic if the data type doesn't match the specific type.

This is because, when the type doesn't match, it returns false as the boolean value, so the panic doesn't occur.

Example Program :

package main

import "fmt"

func main() {

	// create an empty interface
	var a interface{}
	a = "Golang"

	// type assertion
	interfaceValue, booleanValue := a.(int)

	fmt.Println("Interface Value:", interfaceValue)
	fmt.Println("Boolean Value:", booleanValue)
}


Here, you can see we get 0 and false as output because the datattype of values (string) doesn't match with the specified type (int).